Refactor filters and time logic into dedicated modules
- Introduced `filters` and `time` modules for cleaner separation
- Moved `ParseFilters` and `SearchFilters` implementations to `filters`
- Centralized time-related logic into the `time.rs` module
- Replaced redundant `parser_with_filters` function with traits in `filters`
- Simplified dependencies in `monocle.rs` by leveraging modularized functions
